Fixing Nutmeg crashing when axis limits are set in QML
If the axes had no data when the axis limits were set, NaNs popped up
because the default is to set the axis limits to the data limits.
If the data is not valid at the time, stuff gets weird.
Now if the data is not valid, the axis limits because Rect(0,0, 1x1)